灵足关节集成完成

+++ external/soul_link/src/utils.cpp
@@ -4,17 +4,17 @@
 float uint16_to_float(uint16_t x, float x_min, float x_max, int bits) {
   uint32_t span = (1 << bits) - 1;
   float offset = x_max - x_min;
-  return offset * x / span + x_min;
+  return offset * (float)x / (float)span + x_min;
 }
 
-int float_to_uint(float x, float x_min, float x_max, int bits) {
+uint32_t float_to_uint(float x, float x_min, float x_max, int bits) {
   float span = x_max - x_min;
   float offset = x_min;
   if (x > x_max)
     x = x_max;
   else if (x < x_min)
     x = x_min;
-  return (int)((x - offset) * ((float)((1 << bits) - 1)) / span);
+  return (uint32_t)((x - offset) * ((float)((1 << bits) - 1)) / span);
 }
 
 float Byte_to_float(uint8_t *bytedata) {
